+++ ubiProject/src/main/resources/mappers/mytown-borad-mapper.xml
@@ -9,8 +9,8 @@
   SELECT
     B.BOARD_NO,
     B.BOARD_TITLE,
-    SUBSTR(B.BOARD_CONTENT, 1, 100) AS BOARD_CONTENT, -- 앞 100자만
-         <![CDATA[
+ SUBSTR(B.BOARD_CONTENT, 1, 100) AS BOARD_CONTENT,
+  <![CDATA[
   CASE 
     WHEN SYSDATE - B.BOARD_DATE < 1/24/60 THEN FLOOR((SYSDATE - B.BOARD_DATE)*24*60*60) || '초 전'
     WHEN SYSDATE - B.BOARD_DATE < 1/24 THEN FLOOR((SYSDATE - B.BOARD_DATE)*24*60) || '분 전'
@@ -29,59 +29,94 @@ B.BOARD_READ_COUNT,
     M.REGION_CITY,
     M.REGION_DISTRICT,
     
-    
-
-    
-
+   
     (SELECT COUNT(*) FROM BOARD_LIKE L WHERE L.BOARD_NO = B.BOARD_NO) AS LIKE_COUNT,
 
-
+    
     (SELECT IMAGE_PATH FROM BOARD_IMAGE I 
      WHERE I.BOARD_NO = B.BOARD_NO AND I.IMAGE_ORDER = 0 AND ROWNUM = 1) AS THUMBNAIL,
+     
 
 
     (SELECT LISTAGG(H.HASHTAG_NAME, ',') 
      WITHIN GROUP (ORDER BY H.HASHTAG_NAME) 
      FROM BOARD_HASHTAG H 
-     WHERE H.BOARD_NO = B.BOARD_NO) AS HASHTAGS
+     WHERE H.BOARD_NO = B.BOARD_NO) AS HASHTAGS,
      
+        LS.BOARD_TYPE,
+    LS.POST_TYPE
+    
   FROM BOARD B
   JOIN MEMBER M ON B.MEMBER_NO = M.MEMBER_NO
+   JOIN BOARD_TYPE BT ON B.BOARD_TYPE = BT.BOARD_CODE
+  LEFT JOIN LOCAL_SUBCATEGORY LS ON B.BOARD_TYPE = LS.BOARD_TYPE
+  
   WHERE B.BOARD_TYPE = 3
     AND B.BOARD_DEL_FL = 'N'
+    
   ORDER BY B.BOARD_DATE DESC
 </select>
 
 <!-- 상세조회 -->
-  <select id="selectLocalBoardDetail" parameterType="int" resultType="Board">
-    SELECT
-      B.BOARD_NO,
-      B.BOARD_TITLE,
-      B.BOARD_CONTENT,
-      B.BOARD_DATE,
-      B.STAR_COUNT,
-      B.READ_COUNT,
-      M.MEMBER_NICKNAME,
-      M.MEMBER_IMG,
-      M.REGION_CITY,
-      M.REGION_DISTRICT,
+ <select id="selectLocalBoardDetail" parameterType="int" resultType="edu.kh.project.board.model.dto.Board">
+SELECT  
+  B.BOARD_NO,
+  B.BOARD_TITLE,
+  B.BOARD_CONTENT,
+  B.BOARD_DATE,
+  B.BOARD_READ_COUNT,
+  NVL(B.STAR_COUNT, 0) AS STAR_COUNT,
+  
 
-      (SELECT COUNT(*) FROM BOARD_LIKE L WHERE L.BOARD_NO = B.BOARD_NO) AS LIKE_COUNT,
+  NVL((
+    SELECT COUNT(*) 
+    FROM BOARD_LIKE 
+    WHERE BOARD_NO = B.BOARD_NO
+  ), 0) AS LIKE_COUNT,
+  
 
-      (SELECT LISTAGG(I.IMAGE_PATH, ',') 
-       WITHIN GROUP (ORDER BY I.IMAGE_ORDER)
-       FROM BOARD_IMAGE I 
-       WHERE I.BOARD_NO = B.BOARD_NO) AS IMAGE_LIST,
+  M.MEMBER_NICKNAME,
 
-      (SELECT LISTAGG(H.HASHTAG_NAME, ',') 
-       WITHIN GROUP (ORDER BY H.HASHTAG_NAME)
-       FROM BOARD_HASHTAG H 
-       WHERE H.BOARD_NO = B.BOARD_NO) AS HASHTAGS
 
-    FROM BOARD B
-    JOIN MEMBER M ON B.MEMBER_NO = M.MEMBER_NO
-    WHERE B.BOARD_NO = #{boardNo}
-  </select>
+  NVL((
+    SELECT LISTAGG(IMAGE_PATH, ',') WITHIN GROUP (ORDER BY IMAGE_ORDER)
+    FROM BOARD_IMAGE 
+    WHERE BOARD_NO = B.BOARD_NO
+  ), '') AS IMAGE_LIST,
+
+
+  NVL((
+    SELECT LISTAGG(H.HASHTAG_NAME, ',') 
+    FROM BOARD_HASHTAG H
+    WHERE H.BOARD_NO = B.BOARD_NO
+  ), '') AS HASHTAGS,
+
+  BT.BOARD_NAME,
+  M.REGION_CITY,
+  M.REGION_DISTRICT,
+  B.API_SERVICE_ID,
+  B.FACILITY_API_SERVICE_ID,
+
+
+  LS.POST_TYPE
+
+FROM BOARD B
+
+-- 기본 JOIN
+JOIN MEMBER M ON B.MEMBER_NO = M.MEMBER_NO
+JOIN BOARD_TYPE BT ON B.BOARD_TYPE = BT.BOARD_CODE
+
+
+LEFT JOIN (
+  SELECT BOARD_TYPE, MIN(POST_TYPE) AS POST_TYPE
+  FROM LOCAL_SUBCATEGORY
+  GROUP BY BOARD_TYPE
+) LS ON B.BOARD_TYPE = LS.BOARD_TYPE
+
+WHERE B.BOARD_NO = #{boardNo}
+</select>
+
+
